Clean Code

+++ src/Realtor.Service/DTOs/CottageBlockParts/CottageBlockPartCreationDto.cs
@@ -6,9 +6,9 @@ public class CottageBlockPartCreationDto
 {
         public int NumberOfFloors { get; set; }
         public int NumberOfRooms { get; set; }
-        public long TotalArea { get; set; }
+        public int TotalArea { get; set; }
         public AreaUnit TotalAreaUnit { get; set; }
-        public long UsageArea { get; set; }
+        public int UsageArea { get; set; }
         public AreaUnit UsageAreaUnit { get; set; }
         public decimal Cost { get; set; }
         public Currency Currency { get; set; }
